Restructure of frontend files
Current progress on frontend files here. I reshuffled the way the files were organized in order to make them easier to find and alter going forward. I am working on adding visuals/functionality to our registration and home pages respectively but a bug occuring in login.component.ts is currently stopping my progress because whenever I attempt to compile the code I get an error saying the module can't be resolved. When I look on StackOverflow it appears to be in someway linked to paths for the login.component.css file and whether that should be an absolute or relative path. That's as far as I've gotten so far.